+++ src/app/api/user/medias/route.ts
@@ -15,7 +15,17 @@ export async function GET(request: NextApiRequest) {
       return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
-    const user = await User.findOne({ email: token.email }).populate("medias");
+    const searchParams = new URLSearchParams(request.url?.split("?")[1] || "");
+    const searchTerm = searchParams.get("search");
+    let matchQuery = {};
+    if (searchTerm) {
+      matchQuery = { title: { $regex: searchTerm, $options: "i" } };
+    }
+    console.log("match query >", matchQuery);
+    const user = await User.findOne({ email: token.email }).populate({
+      path: "medias",
+      match: matchQuery,
+    });
     if (!user) {
       return NextResponse.json({ error: "User not found" }, { status: 404 });
     }
